Crate hopr_primitive_types

This crate contains basic types used throughout the entire HOPR codebase. Types from this crate are not necessarily specific only to HOPR.

Modules

bounded

Contains various size-bounded types

errors

Enumerates all errors in this crate.

prelude

primitives

Implements the most primitive types, such as U256 or Address.

rlp

Obsolete RLP encoding module to be removed in future versions.

sma

Contains various implementations of Simple Moving Average.

traits

Defines commonly used traits across the entire code base.

Functions

f64_approx_eq

Approximately compares two double-precision floats.