Trait hopr_primitive_types::traits::AutoBinarySerializable

pub trait AutoBinarySerializable: Serialize + for<'a> Deserialize<'a> {
    const SIZE: usize = 1usize;
}

Type implementing this trait has automatic binary serialization/deserialization capability using the default binary format, which is currently bincode.

Provided Associated Constants

const SIZE: usize = 1usize

Minimum size of an automatically serialized type in bytes is 1.

Object Safety

This trait is not object safe.