Trait hopr_primitive_types::traits::BinarySerializable

pub trait BinarySerializable: Sized {
    const SIZE: usize;

    // Required methods
    fn from_bytes(data: &[u8]) -> Result<Self>;
    fn to_bytes(&self) -> Box<[u8]>;
}

A type that can be serialized and deserialized to a binary form.

Implementing this trait automatically implements ToHex trait which then uses the serialize method.

Required Associated Constants

const SIZE: usize

Minimum size of this type in bytes.

Required Methods

fn from_bytes(data: &[u8]) -> Result<Self>

Deserializes the type from a binary blob.

fn to_bytes(&self) -> Box<[u8]>

Serializes the type into a fixed size binary blob.

Object Safety

This trait is not object safe.