Crate hopr_primitive_types

This crate contains basic types used throughout the entire HOPR codebase. Types from this crate are not necessarily specific only to HOPR.

Modules

balance

Contains implementations for the token and coin balance types.

bounded

Contains various size-bounded types

errors

Lists all errors in this crate.

prelude

primitives

Implements the most primitive types, such as U256 or Address.

sma

Contains various implementations of Simple Moving Average.

traits

Defines commonly used traits across the entire code base.

Functions

f64_approx_eq

Approximately compares two double-precision floats.